use sopel core where we can (#138) (#155)
* use sopel core where we can

* move botver to responses (#139)

* Delete pyver.py

.version does this fine
