+++ modules/responses.py
@@ -27,3 +27,17 @@ def cancel(bot, trigger):
     """Cancel reminder."""
     bot.reply(('Pinging RhinosF1, Reception123, Voidwalker or Zppix to cancel '
                '{}\'s reminder.').format(trigger.nick))
+
+
+@commands('botversion', 'bv')
+@example('.botversion')
+def botversion(bot, trigger):
+    """List the current version of the bot."""
+    bot.say('The current version of this bot is 6.0 (v6)')
+
+
+@commands('source', 'botsource')
+@example('.source')
+def githubsource(bot, trigger):
+    """Give the link to ZppixBot's Github."""
+    bot.reply('My code can be found here: https://github.com/Pix1234/ZppixBot-Source')
